:sparkles: Scala 3.3.7対応

+++ core/src/main/scala/io/github/ondeoma/scalactive/controllers/ComponentController.scala
@@ -27,7 +27,7 @@ trait ComponentController[CC <: ComponentController[?]] {
       clear()
       init()
       restoreScrollPosition()
-    }, ScalactiveConfig.reloadTimerLazyDuration.toMillis)
+    }, ScalactiveConfig.reloadTimerLazyDuration.toMillis.toDouble)
   }
 
   def backupScrollPosition(): Unit = {
